\name{addTo}
\alias{addTo}
\title{Add child elements to a parent element.}
\usage{
  addTo(parent, ..., row = NA, column = NA)
}
\arguments{
  \item{parent}{Parent element.}

  \item{...}{One or more child elements.}

  \item{row}{If parent element is a table row and column
  indices must be provided to add supplementary results to
  cell in the table.}

  \item{column}{If parent element is a table row and column
  indices must be provided to add supplementary results to
  cell in the table.}
}
\value{
  Updated parent element.
}
\description{
  Add child elements to a parent element.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

