\name{addToInput}
\alias{addToInput}
\title{Add elements to the "Input" subsection in the "Methods & Data" section of a standard report.}
\usage{
  addToInput(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{Elements that will be added to the input
  section. Often these are paragraph elements.}
}
\value{
  Updated report element.
}
\description{
  Add elements to the "Input" subsection in the "Methods &
  Data" section of a standard report.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

