\name{addToIntroduction}
\alias{addToIntroduction}
\title{Add elements to the "Introduction" subsection in the "Overview" section of a standard report.}
\usage{
  addToIntroduction(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{Elements that will be added to the
  introduction section. Often these are paragraph
  elements.}
}
\value{
  Updated report element.
}
\description{
  Add elements to the "Introduction" subsection in the
  "Overview" section of a standard report.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

