\name{addToMeta}
\alias{addToMeta}
\title{Add elements to the "Meta" section of a standard report.}
\usage{
  addToMeta(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{Elements that will be added to the references
  section. These may be any elements.}
}
\value{
  Updated report element.
}
\description{
  Add elements to the "Meta" section of a standard report.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

