\name{addToMethods}
\alias{addToMethods}
\title{Add elements to the "Methds & Data" section of a standard report.}
\usage{
  addToMethods(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{Elements that will be added to the methods
  section. Often these are subsection elements.}
}
\value{
  Updated report element.
}
\description{
  Add elements to the "Methds & Data" section of a standard
  report.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

