\name{addToReferences}
\alias{addToReferences}
\title{Add elements to the "References" subsection in the "Methods & Data" section of a standard report.}
\usage{
  addToReferences(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{Elements that will be added to the references
  section. These should be citation elements.}
}
\value{
  Updated report element.
}
\description{
  Add elements to the "References" subsection in the
  "Methods & Data" section of a standard report.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

