\name{addToResults}
\alias{addToResults}
\title{Add elements to the "Results" section of a standard report.}
\usage{
  addToResults(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{Elements that will be added to the results
  section. Often these are subsection elements.}
}
\value{
  Updated report element.
}
\description{
  Add elements to the "Results" section of a standard
  report.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

