\name{addToSummary}
\alias{addToSummary}
\title{Add elements to the "Summary" subsection in the "Overview" section of a standard report.}
\usage{
  addToSummary(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{Elements that will be added to the overview
  section. Often these are paragraph elements.}
}
\value{
  Updated report element.
}
\description{
  Add elements to the "Summary" subsection in the
  "Overview" section of a standard report.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

