\name{asCode}
\alias{asCode}

\title{Format source code using a fixed-width font.}

\description{Formats a string using a fixed-width font.}

\usage{asCode(...)}

\arguments{
\item{\dots}{Strings and scalar values to be contatenated and formatted.}
}
\details{}

\value{Returns an HTML-formatted string of the concatenated arguments.}
\author{Nils Gehlenborg (nils@hms.harvard.edu)}
