\name{asEmph}
\alias{asEmph}
\title{Format text with emphasis (usually resulting in text set in italics).}
\usage{
  asEmph(...)
}
\arguments{
  \item{...}{One or more strings that will be
  concatenated.}
}
\value{
  Text with emphasis.
}
\description{
  Format text with emphasis (usually resulting in text set
  in italics).
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

