\name{asFilename}
\alias{asFilename}

\title{Format a filename using a fixed-width font.}

\description{Formats a string using a fixed-width font.}

\usage{asFilename(...)}

\arguments{
\item{\dots}{Strings and scalar values to be contatenated and formatted.}
}
\details{}

\value{Returns an HTML-formatted string of the concatenated arguments.}
\author{Nils Gehlenborg (nils@hms.harvard.edu)}
