\name{asLink}
\alias{asLink}
\title{Format text as a hyperlink.}
\usage{
  asLink(url, ...)
}
\arguments{
  \item{url}{URL to be for the link.}

  \item{...}{One or more strings that will be concatenated
  to form the text of the link.}
}
\value{
  A hyperlink.
}
\description{
  Format text as a hyperlink.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

