\name{asParameter}
\alias{asParameter}
\title{Format text as parameter.}
\usage{
  asParameter(...)
}
\arguments{
  \item{...}{One or more strings that will be
  concatenated.}
}
\value{
  Text formatted as a parameter.
}
\description{
  Format text as parameter.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

