\name{asReference}
\alias{asReference}
\title{Reference a citation, figure or table element.}
\usage{
  asReference(element)
}
\arguments{
  \item{element}{Citation, figure or table element.}
}
\value{
  A reference string for the referenced element that will
  be resolved when the report is written to file and
  rendered to HTML.
}
\description{
  Reference a citation, figure or table element.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

