\name{asStrong}
\alias{asStrong}
\title{Format text with strong emphasis (usually resulting in text set in bold).}
\usage{
  asStrong(...)
}
\arguments{
  \item{...}{One or more strings that will be
  concatenated.}
}
\value{
  Text with strong emphasis.
}
\description{
  Format text with strong emphasis (usually resulting in
  text set in bold).
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

