\name{asSummary}
\alias{asSummary}
\title{Include a result in text. This is a legacy method and provided only for backwards compatibility.}
\usage{
  asSummary(result)
}
\arguments{
  \item{result}{The result element.}
}
\value{
  The result element.
}
\description{
  Include a result in text. This is a legacy method and
  provided only for backwards compatibility.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

