\name{asValue}
\alias{asValue}
\title{Format text as value.}
\usage{
  asValue(...)
}
\arguments{
  \item{...}{One or more strings that will be
  concatenated.}
}
\value{
  Text formatted as a value.
}
\description{
  Format text as value.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

