\docType{data}
\name{DEFAULT.SIGNIFICANT.ENTITY}
\alias{DEFAULT.SIGNIFICANT.ENTITY}
\title{Name of entities that are labeled as signficiant.}
\format{chr "statistically significant findings"}
\usage{
  DEFAULT.SIGNIFICANT.ENTITY
}
\description{
  Name of entities that are labeled as signficiant.
}
\keyword{datasets}

