\name{getCollectionDate}
\alias{getCollectionDate}
\title{Get the collection date of \code{report}. The collection date is the date when the collection that this report is part of was created.}
\usage{
  getCollectionDate(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Collection date of \code{report} (a string)
}
\description{
  Get the collection date of \code{report}. The collection
  date is the date when the collection that this report is
  part of was created.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

