\name{getCollectionVersion}
\alias{getCollectionVersion}
\title{Get the collection version of \code{report}. The collection version is the version string of the collection that this report is part of.}
\usage{
  getCollectionVersion(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Collection version of \code{report} (a string)
}
\description{
  Get the collection version of \code{report}. The
  collection version is the version string of the
  collection that this report is part of.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

