\name{getContactInformationEmail}
\alias{getContactInformationEmail}
\title{Get contact email address for \code{report}.}
\usage{
  getContactInformationEmail(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Contact email address.
}
\description{
  Get contact email address for \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

