\name{getContactInformationLabel}
\alias{getContactInformationLabel}
\title{Get label for contact button for \code{report}.}
\usage{
  getContactInformationLabel(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Contact email button label.
}
\description{
  Get label for contact button for \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

