\name{getContactInformationMessage}
\alias{getContactInformationMessage}
\title{Get contact email default message for \code{report}.}
\usage{
  getContactInformationMessage(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Contact email default message.
}
\description{
  Get contact email default message for \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

