\name{getContactInformationSubject}
\alias{getContactInformationSubject}
\title{Get contact email subject line for \code{report}.}
\usage{
  getContactInformationSubject(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Contact email subject line.
}
\description{
  Get contact email subject line for \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

