\name{getCopyrightOwner}
\alias{getCopyrightOwner}
\title{Get name of the copyright owner for \code{report}.}
\usage{
  getCopyrightOwner(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Name of the copyright owner.
}
\description{
  Get name of the copyright owner for \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

