\name{getCopyrightStatement}
\alias{getCopyrightStatement}
\title{Get copyright statement for \code{report}. This text is linked to the copyright URL.}
\usage{
  getCopyrightStatement(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Text of the c opyright statement.
}
\description{
  Get copyright statement for \code{report}. This text is
  linked to the copyright URL.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

