\name{getCopyrightUrl}
\alias{getCopyrightUrl}
\title{Get copyright URL for \code{report}, which is linked to the copyright statement.}
\usage{
  getCopyrightUrl(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Copyright URL.
}
\description{
  Get copyright URL for \code{report}, which is linked to
  the copyright statement.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

