\name{getCopyrightYear}
\alias{getCopyrightYear}
\title{Get copyright year \code{report}.}
\usage{
  getCopyrightYear(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Copyright year.
}
\description{
  Get copyright year \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

