\name{getCreatorDate}
\alias{getCreatorDate}
\title{Get date when \code{report} was created.}
\usage{
  getCreatorDate(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Date when \code{report} was created or NA.
}
\description{
  Get date when \code{report} was created.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

