\name{getCreatorName}
\alias{getCreatorName}
\title{Get name and version of the Nozzle package that was used to create \code{report}.}
\usage{
  getCreatorName(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Name and version of the Nozzle package that created
  \code{report} or NA.
}
\description{
  Get name and version of the Nozzle package that was used
  to create \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

