\name{getCustomPrintCss}
\alias{getCustomPrintCss}
\title{Get the path or URL of the CSS file to be used to overwrite the default print (not: screen) style sheet.}
\usage{
  getCustomPrintCss(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Path or URL of CSS file. Can be relative or absolute.
}
\description{
  Get the path or URL of the CSS file to be used to
  overwrite the default print (not: screen) style sheet.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

