\name{getCustomScreenCss}
\alias{getCustomScreenCss}
\title{Get the path or URL of the CSS file to be used to overwrite the default screen (not: print) style sheet.}
\usage{
  getCustomScreenCss(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Path or URL of CSS file. Can be relative or absolute.
}
\description{
  Get the path or URL of the CSS file to be used to
  overwrite the default screen (not: print) style sheet.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

