\name{getDoi}
\alias{getDoi}
\title{Get the DOI (document object identifier, http://www.doi.org) for \code{report}.}
\usage{
  getDoi(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Document object identifier (DOI) for \code{report}.
}
\description{
  Get the DOI (document object identifier,
  http://www.doi.org) for \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

