\name{getDoiCreator}
\alias{getDoiCreator}
\title{Get the DOI creator for \code{report}.}
\usage{
  getDoiCreator(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Creator associated with the DOI of the report
  \code{report}.
}
\description{
  Get the DOI creator for \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

