\name{getDoiPublisher}
\alias{getDoiPublisher}
\title{Get the DOI publisher for \code{report}.}
\usage{
  getDoiPublisher(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Publisher associated with the DOI of the report
  \code{report}.
}
\description{
  Get the DOI publisher for \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

