\name{getDoiResolver}
\alias{getDoiResolver}
\title{Get the DOI resolver URL (e.g. http://dx.doi.org) for \code{report}.}
\usage{
  getDoiResolver(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Document object identifier (DOI) resolver URL for
  \code{report}.
}
\description{
  Get the DOI resolver URL (e.g. http://dx.doi.org) for
  \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

