\name{getDoiTitle}
\alias{getDoiTitle}
\title{Get the DOI title for \code{report}.}
\usage{
  getDoiTitle(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Title associated with the DOI of the report
  \code{report}.
}
\description{
  Get the DOI title for \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

