\name{getDoiVersion}
\alias{getDoiVersion}
\title{Get the DOI version for \code{report}.}
\usage{
  getDoiVersion(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Version associated with the DOI of the report
  \code{report}.
}
\description{
  Get the DOI version for \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

