\name{getDoiYear}
\alias{getDoiYear}
\title{Get the DOI year for \code{report}.}
\usage{
  getDoiYear(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Publication year associated with the DOI of the report
  \code{report}.
}
\description{
  Get the DOI year for \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

