\name{getExportedElementIds}
\alias{getExportedElementIds}
\title{Get the IDs of exported elements from \code{report}. This is an experimental feature of Nozzle and may not lead to the expected results.}
\usage{
  getExportedElementIds(report)
}
\arguments{
  \item{report}{The source report.}
}
\value{
  The IDs of exported report elements or NULL.
}
\description{
  Get the IDs of exported elements from \code{report}. This
  is an experimental feature of Nozzle and may not lead to
  the expected results.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

