\name{getFigureFile}
\alias{getFigureFile}
\title{Get path or URL of image file associated with a figure element.}
\usage{
  getFigureFile(element)
}
\arguments{
  \item{element}{Figure element.}
}
\value{
  Path or URL of the image file or NA if \code{element} is
  not a figure.
}
\description{
  Get path or URL of image file associated with a figure
  element.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

