\name{getFigureFileHighRes}
\alias{getFigureFileHighRes}
\title{Get path or URL of high-resolution of vector-based image file associated with a figure element.}
\usage{
  getFigureFileHighRes(element)
}
\arguments{
  \item{element}{Figure element.}
}
\value{
  Path or URL of the image file or NA if \code{element} is
  not a figure.
}
\description{
  Get path or URL of high-resolution of vector-based image
  file associated with a figure element.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

