\name{getGoogleAnalyticsId}
\alias{getGoogleAnalyticsId}
\title{Get Google Analytics tracking ID for \code{report}.}
\usage{
  getGoogleAnalyticsId(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Google Analytics Tracking ID or NA if not set.
}
\description{
  Get Google Analytics tracking ID for \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

