\name{getLogo}
\alias{getLogo}
\title{Get logo file for one of six positions (three at the top, three at the bottom) in \code{report}.}
\usage{
  getLogo(report, position)
}
\arguments{
  \item{report}{Report element.}

  \item{position}{One of LOGO.TOP.LEFT, LOGO.TOP.CENTER,
  LOGO.TOP.RIGHT, LOGO.BOTTOM.LEFT, LOGO.BOTTOM.CENTER,
  LOGO.BOTTOM.RIGHT.}
}
\value{
  Path or URL to the logo file at \code{position}.
}
\description{
  Get logo file for one of six positions (three at the top,
  three at the bottom) in \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

