\name{getMaintainerAffiliation}
\alias{getMaintainerAffiliation}
\title{Get affiliation of maintainer of \code{report}.}
\usage{
  getMaintainerAffiliation(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Affiliation of the maintainer.
}
\description{
  Get affiliation of maintainer of \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

