\name{getMaintainerEmail}
\alias{getMaintainerEmail}
\title{Get email address of maintainer of \code{report}.}
\usage{
  getMaintainerEmail(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Email address of the maintainer.
}
\description{
  Get email address of maintainer of \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

