\name{getMaintainerName}
\alias{getMaintainerName}
\title{Get name of maintainer of \code{report}.}
\usage{
  getMaintainerName(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Name of the maintainer.
}
\description{
  Get name of maintainer of \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

