\name{getRendererDate}
\alias{getRendererDate}
\title{Get date when \code{report} was rendered.}
\usage{
  getRendererDate(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Date when \code{report} was rendered or NA.
}
\description{
  Get date when \code{report} was rendered.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

