\name{getRendererName}
\alias{getRendererName}
\title{Get name and version of the Nozzle package that was used to render \code{report}.}
\usage{
  getRendererName(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Name and version of the Nozzle package that rendered
  \code{report} or NA.
}
\description{
  Get name and version of the Nozzle package that was used
  to render \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

