\name{getReportId}
\alias{getReportId}
\title{Get the ID (a UUID) of \code{report}.}
\usage{
  getReportId(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  ID of \code{report} or NA.
}
\description{
  Get the ID (a UUID) of \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

