\name{getReportSubTitle}
\alias{getReportSubTitle}
\title{Get the subtitle of \code{report}.}
\usage{
  getReportSubTitle(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  SubTitle of \code{report}
}
\description{
  Get the subtitle of \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

