\name{getReportTitle}
\alias{getReportTitle}
\title{Get the title of \code{report}.}
\usage{
  getReportTitle(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Title of \code{report}
}
\description{
  Get the title of \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

