\name{getSignificantEntity}
\alias{getSignificantEntity}
\title{Get name of entities that are called out as significant, e.g. "gene". This is currently not being used and might become obsolete in future versions of Nozzle.}
\usage{
  getSignificantEntity(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Name of entities called out as significant.
}
\description{
  Get name of entities that are called out as significant,
  e.g. "gene". This is currently not being used and might
  become obsolete in future versions of Nozzle.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

