\name{getSignificantResultsCount}
\alias{getSignificantResultsCount}
\title{Get the total number of significant results in \code{report}.}
\usage{
  getSignificantResultsCount(report)
}
\arguments{
  \item{report}{The report.}
}
\value{
  Number of signficiant results.
}
\description{
  Get the total number of significant results in
  \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

