\name{getSoftwareName}
\alias{getSoftwareName}
\title{Get the name of the software that used Nozzle to generate \code{report}.}
\usage{
  getSoftwareName(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Name of the software that used Nozzle to generate
  \code{report}.
}
\description{
  Get the name of the software that used Nozzle to generate
  \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

