\name{getSoftwareVersion}
\alias{getSoftwareVersion}
\title{Get the version of the software that used Nozzle to generate \code{report}.}
\usage{
  getSoftwareVersion(report)
}
\arguments{
  \item{report}{Report element.}
}
\value{
  Version of the software that used Nozzle to generate
  \code{report}.
}
\description{
  Get the version of the software that used Nozzle to
  generate \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

