\name{getSummary}
\alias{getSummary}
\title{Get the first element of the "Summary" subsection in the "Overview" section in a standard report.}
\usage{
  getSummary(report)
}
\arguments{
  \item{report}{The report.}
}
\value{
  A report element.
}
\description{
  Get the first element of the "Summary" subsection in the
  "Overview" section in a standard report.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

