\name{getTableFile}
\alias{getTableFile}
\title{Get path or URL of file associatd with table element.}
\usage{
  getTableFile(element)
}
\arguments{
  \item{element}{Table element.}
}
\value{
  Path or URL.
}
\description{
  Get path or URL of file associatd with table element.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

