\name{isFigure}
\alias{isFigure}
\title{Test if \code{element} is a figure element.}
\usage{
  isFigure(element)
}
\arguments{
  \item{element}{Element to test.}
}
\value{
  TRUE if the element is a figure, FALSE otherwise.
}
\description{
  Test if \code{element} is a figure element.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

