\name{isTable}
\alias{isTable}
\title{Test if \code{element} is a table element.}
\usage{
  isTable(element)
}
\arguments{
  \item{element}{Element to test.}
}
\value{
  TRUE if the element is a table, FALSE otherwise.
}
\description{
  Test if \code{element} is a table element.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

