\name{newCustomReport}
\alias{newCustomReport}
\title{Create a new custom report without pre-defined sections.}
\usage{
  newCustomReport(..., version = 0)
}
\arguments{
  \item{...}{One or more strings that will be concatenated
  into the report title.}

  \item{version}{Version number. Not in use.}
}
\value{
  A new report element.
}
\description{
  Create a new custom report without pre-defined sections.
}
\author{
  Nils Gehlenborg (nils@hms.harvard.edu)
}

