\name{newReport}
\alias{newReport}
\title{Create a new report with pre-defined sections Overview/Introduction, Overview/Summary, Results, Methods & Data/Input, Methods & Data/References and Meta Data.}
\usage{
  newReport(..., version = 0)
}
\arguments{
  \item{...}{One or more strings that will be concatenated
  into the report title.}

  \item{version}{Version number. Not in use.}
}
\value{
  A new report element.
}
\description{
  Create a new report with pre-defined sections
  Overview/Introduction, Overview/Summary, Results, Methods
  & Data/Input, Methods & Data/References and Meta Data.
}
\author{
  Nils Gehlenborg (nils@hms.harvard.edu)
}

