\name{newWebCitation}
\alias{newWebCitation}
\title{Create a citation element that represents a document published online. This is a convenience wrapper for \code{newCitation}.}
\usage{
  newWebCitation(authors, title, url)
}
\arguments{
  \item{authors}{Names of authors.}

  \item{title}{Title of the document.}

  \item{url}{URL of the document.}
}
\value{
  New element.
}
\description{
  Create a citation element that represents a document
  published online. This is a convenience wrapper for
  \code{newCitation}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

