\name{Nozzle-v1.beta-package}
\alias{Nozzle-v1.beta-package}
\alias{Nozzle-v1.beta}
\docType{package}
\title{What the package does (short line)}
\description{More about what it does (maybe more than one line)}
\details{
\tabular{ll}{
Package: \tab Nozzle-v1.beta\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-02-16\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
An overview of how to use the package, including the most important functions
}
\author{
Nils Gehlenborg, Harvard Medical School
Maintainer: Nils Gehlenborg <nils@hms.harvard.edu>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
