\docType{package}
\name{Nozzle.R1-package}
\alias{Nozzle.R1-package}
\alias{nozzle}
\title{Nozzle: a Report Generation Toolkit for Data Analysis Pipelines}
\description{
  Provides a high-level API to generate HTML reports with
  dynamic user interface elements.
}
\details{
  \tabular{ll}{ Package: \tab Nozzle.R1\cr Type: \tab
  Package\cr Version: \tab 0.3-1\cr Date: \tab
  2013-01-03\cr License: \tab BSD (>= 2)\cr LazyLoad: \tab
  no\cr }

  Nozzle was designed to facilitate summarization and rapid
  browsing of complex results in data analysis pipelines
  where multiple analyses are performed frequently on big
  data sets.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}
\references{
  \url{http://gdac.broadinstitute.org}
  \url{http://www.github.com/parklab/Nozzle}
}
\keyword{package}

