\name{setCopyright}
\alias{setCopyright}
\title{Set copyright messsage for \code{report}.}
\usage{
  setCopyright(report, owner, year, statement = NA,
    url = NA)
}
\arguments{
  \item{report}{Report element.}

  \item{owner}{Copyright owner, e.g. "The President and
  Fellows of Harvard College".}

  \item{year}{Copyright year, e.g. "2012" or "2011-2013".}

  \item{statement}{Copyright statement, e.g. "All rights
  reserved.".}

  \item{url}{A URL that will be linked to the copyright
  owner.}
}
\value{
  Updated report element.
}
\description{
  Set copyright messsage for \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

