\name{setCustomPrintCss}
\alias{setCustomPrintCss}
\title{Set the path or URL of the CSS file to be used to overwrite the default print (not: screen) style sheet. Can be relative or absolute.}
\usage{
  setCustomPrintCss(report, cssFile)
}
\arguments{
  \item{report}{Report element.}

  \item{cssFile}{URL or a relative or absolute path to a
  CSS file.}
}
\value{
  Updated report element.
}
\description{
  Set the path or URL of the CSS file to be used to
  overwrite the default print (not: screen) style sheet.
  Can be relative or absolute.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

