\name{setDoiCreator}
\alias{setDoiCreator}
\title{Set the DOI meta data creator for \code{report}.}
\usage{
  setDoiCreator(report, creator)
}
\arguments{
  \item{report}{Report element.}

  \item{creator}{Name of the report creator/author.}
}
\value{
  Updated report element.
}
\description{
  Set the DOI meta data creator for \code{report}.
}
\note{
  This should match the meta data stored for the DOI.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

