\name{setDoiPublisher}
\alias{setDoiPublisher}
\title{Set the DOI meta data publisher for \code{report}.}
\usage{
  setDoiPublisher(report, publisher)
}
\arguments{
  \item{report}{Report element.}

  \item{publisher}{Publisher of the report.}
}
\value{
  Updated report element.
}
\description{
  Set the DOI meta data publisher for \code{report}.
}
\note{
  This should match the meta data stored for the DOI.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

