\name{setDoiTitle}
\alias{setDoiTitle}
\title{Set the DOI meta data title for \code{report}.}
\usage{
  setDoiTitle(report, title)
}
\arguments{
  \item{report}{Report element.}

  \item{title}{Title of the report.}
}
\value{
  Updated report element.
}
\description{
  Set the DOI meta data title for \code{report}.
}
\note{
  This should match the meta data stored for the DOI.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

