\name{setDoiVersion}
\alias{setDoiVersion}
\title{Set the DOI meta data version for \code{report}.}
\usage{
  setDoiVersion(report, version)
}
\arguments{
  \item{report}{Report element.}

  \item{version}{Version of the report.}
}
\value{
  Updated report element.
}
\description{
  Set the DOI meta data version for \code{report}.
}
\note{
  This should match the meta data stored for the DOI.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

