\name{setDoiYear}
\alias{setDoiYear}
\title{Set the DOI meta data year for \code{report}.}
\usage{
  setDoiYear(report, year)
}
\arguments{
  \item{report}{Report element.}

  \item{year}{Publication year of the report.}
}
\value{
  Updated report element.
}
\description{
  Set the DOI meta data year for \code{report}.
}
\note{
  This should match the meta data stored for the DOI.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

