\name{setFigureFile}
\alias{setFigureFile}
\title{Set path or URL of image file associated with a figure element. Paths can relative or absolute.}
\usage{
  setFigureFile(element, file)
}
\arguments{
  \item{element}{A figure element.}

  \item{file}{Path or URL of the image file.}
}
\value{
  Updated figure element or NA is \code{element} is not a
  figure.
}
\description{
  Set path or URL of image file associated with a figure
  element. Paths can relative or absolute.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

