\name{setFigureFileHighRes}
\alias{setFigureFileHighRes}
\title{Set path or URL of high-resolution or vector-based image file associated with a figure element. Paths can be relative or absolute.}
\usage{
  setFigureFileHighRes(element, file)
}
\arguments{
  \item{element}{A figure element.}

  \item{file}{Path or URL of the image file.}
}
\value{
  Updated figure element or NA is \code{element} is not a
  figure.
}
\description{
  Set path or URL of high-resolution or vector-based image
  file associated with a figure element. Paths can be
  relative or absolute.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

