\name{setGoogleAnalyticsId}
\alias{setGoogleAnalyticsId}
\title{Set the Google Analytics tracking ID to be embedded in this report ("web property id", usually starts with "UA-").}
\usage{
  setGoogleAnalyticsId(report, id)
}
\arguments{
  \item{report}{Report element.}

  \item{id}{Web property ID for the Google Analytics
  tracking account.}
}
\value{
  Updated report element.
}
\description{
  Set the Google Analytics tracking ID to be embedded in
  this report ("web property id", usually starts with
  "UA-").
}
\note{
  A (free) Google Analytics account is required.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}
\references{
  \url{http://www.google.com/analytics}
}

