\name{setMaintainerAffiliation}
\alias{setMaintainerAffiliation}
\title{Set affiliation of maintainer of \code{report}.}
\usage{
  setMaintainerAffiliation(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{One or more strings that will be concatenated
  to form the affiliation of the maintainer.}
}
\value{
  Updated report element.
}
\description{
  Set affiliation of maintainer of \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

