\name{setMaintainerEmail}
\alias{setMaintainerEmail}
\title{Set email address of maintainer of \code{report}.}
\usage{
  setMaintainerEmail(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{One or more strings that will be concatenated
  to form the email address of the maintainer.}
}
\value{
  Updated report element.
}
\description{
  Set email address of maintainer of \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

