\name{setMaintainerName}
\alias{setMaintainerName}
\title{Set name of maintainer of \code{report}.}
\usage{
  setMaintainerName(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{One or more strings that will be concatenated
  to form the name of the maintainer.}
}
\value{
  Updated report element.
}
\description{
  Set name of maintainer of \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

