\name{setNextReport}
\alias{setNextReport}
\title{Set the URL and title of the "next" report after \code{report}. This will be accessible through the utility menu.}
\usage{
  setNextReport(report, url, ...)
}
\arguments{
  \item{report}{Report object.}

  \item{url}{URL of the next report.}

  \item{...}{One or more strings that will be concatenated
  to form the title of the next report.}
}
\value{
  Updated report element.
}
\description{
  Set the URL and title of the "next" report after
  \code{report}. This will be accessible through the
  utility menu.
}
\author{
  nils
}

