\name{setParentReport}
\alias{setParentReport}
\title{Set the URL and title of the "parent" report above \code{report}. This will be accessible through the utility menu.}
\usage{
  setParentReport(report, url, ...)
}
\arguments{
  \item{report}{Report object.}

  \item{url}{URL of the next report.}

  \item{...}{One or more strings that will be concatenated
  to form the title of the parent report.}
}
\value{
  Updated report element.
}
\description{
  Set the URL and title of the "parent" report above
  \code{report}. This will be accessible through the
  utility menu.
}
\author{
  nils
}

