\name{setPreviousReport}
\alias{setPreviousReport}
\title{Set the URL and title of the "previous" report before \code{report}. This will be accessible through the utility menu.}
\usage{
  setPreviousReport(report, url, ...)
}
\arguments{
  \item{report}{Report object.}

  \item{url}{URL of the next report.}

  \item{...}{One or more strings that will be concatenated
  to form the title of the previous report.}
}
\value{
  Updated report element.
}
\description{
  Set the URL and title of the "previous" report before
  \code{report}. This will be accessible through the
  utility menu.
}
\author{
  nils
}

