\name{setReportSubTitle}
\alias{setReportSubTitle}
\title{Set the subtitle of \code{report}.}
\usage{
  setReportSubTitle(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{One or more strings that will be concatenated
  to form the subtitle of the report.}
}
\value{
  Updated report element or NA if \code{report} has no
  subtitle element.
}
\description{
  Set the subtitle of \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

