\name{setReportTitle}
\alias{setReportTitle}
\title{Set the title of \code{report}.}
\usage{
  setReportTitle(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{One or more strings that will be concatenated
  to form the title of the report.}
}
\value{
  Updated report element or NA if \code{report} has no
  title element.
}
\description{
  Set the title of \code{report}.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

