\name{setSignificantEntity}
\alias{setSignificantEntity}
\title{Set name of entities that are called out as significant, e.g. "gene". This is currently not being used and might become obsolete in future versions of Nozzle.}
\usage{
  setSignificantEntity(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{List of strings that will be concatenated to
  form the name of the entities.}
}
\value{
  Updated report element.
}
\description{
  Set name of entities that are called out as significant,
  e.g. "gene". This is currently not being used and might
  become obsolete in future versions of Nozzle.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

