\name{setSoftwareName}
\alias{setSoftwareName}
\title{Set the name of the software that used Nozzle to generate \code{report}, e.g. "My Report Generator Script".}
\usage{
  setSoftwareName(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{One or more strings that will be concatenated
  to form the name of the software.}
}
\value{
  Updated report element.
}
\description{
  Set the name of the software that used Nozzle to generate
  \code{report}, e.g. "My Report Generator Script".
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

