\name{setSoftwareVersion}
\alias{setSoftwareVersion}
\title{Set the name of the software that used Nozzle to generate \code{report}, e.g. "Version 1.2".}
\usage{
  setSoftwareVersion(report, ...)
}
\arguments{
  \item{report}{Report element.}

  \item{...}{One or more strings that will be concatenated
  to form the version of the software.}
}
\value{
  Updated report element.
}
\description{
  Set the name of the software that used Nozzle to generate
  \code{report}, e.g. "Version 1.2".
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

