\name{setTableFile}
\alias{setTableFile}
\title{Set path or URL of file associatd with table element.}
\usage{
  setTableFile(element, file)
}
\arguments{
  \item{element}{Table element.}

  \item{file}{Path or URL to file.}
}
\value{
  Updated element.
}
\description{
  Set path or URL of file associatd with table element.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

