\docType{data}
\name{TABLE.SIGNIFICANT.DIGITS}
\alias{TABLE.SIGNIFICANT.DIGITS}
\title{Default number of significant digits to be used to trim numeric columns in tables.}
\format{num 2}
\usage{
  TABLE.SIGNIFICANT.DIGITS
}
\description{
  Default number of significant digits to be used to trim
  numeric columns in tables.
}
\keyword{datasets}

