\name{asFilename}
\alias{asFilename}
\title{Format text as filename.}
\usage{
  asFilename(...)
}
\arguments{
  \item{...}{One or more strings that will be
  concatenated.}
}
\value{
  Text formatted as a filename.
}
\description{
  Format text as filename.
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}

