\docType{package}
\name{Nozzle.R1-package}
\alias{nozzle}
\alias{Nozzle.R1-package}
\title{Nozzle: a Report Generation Toolkit for Data Analysis Pipelines}
\description{
  Provides a high-level API to generate HTML reports with
  dynamic user interface elements.
}
\details{
  \tabular{ll}{ Package: \tab Nozzle.R1\cr Type: \tab
  Package\cr Version: \tab 1.1-0\cr Date: \tab
  2013-04-16\cr License: \tab LGPL (>= 2)\cr LazyLoad: \tab
  yes\cr }

  Nozzle was designed to facilitate summarization and rapid
  browsing of complex results in data analysis pipelines
  where multiple analyses are performed frequently on big
  data sets.
}
\note{
  The "R1" in the "Nozzle.R1" package name stands for
  "revision 1" of the Nozzle R API. All versions of the
  Nozzle.R1 package will be backwards-compatible and able
  to render reports generated with earlier versions of the
  package. When backwards-compatibility of the API can no
  longer maintained the package name will change to
  "Nozzle.R2".
}
\author{
  Nils Gehlenborg \email{nils@hms.harvard.edu}
}
\references{
  \url{http://www.github.com/parklab/Nozzle},
  \url{http://gdac.broadinstitute.org/nozzle}
}
\keyword{package}

